Refactor mHC Pallas kernel: split forward and backward passes
Split mhc_kernel.py into three files:
- mhc_common.py: Core shared mathematical functions and custom VJP coeffs.
- mhc_fwd_kernels.py: Forward Pallas kernels, sharded wrappers, and fwd VJP.
- mhc_bwd_kernels.py: Backward Pallas kernels, sharded wrappers, and bwd VJP.

This improves maintainability and satisfies formatting/linting constraints.

CHANGE_STEWARD=true
TAG=agy
CONV=19f0bb3e-97df-4163-8aca-a7daa4dbd9cb
